requirements.txt
torch
torchvision
transformers
pandas
tqdm
ftfy
regex
Pillow
git+https://github.com/openai/CLIP.git